pyproject.toml
[project]
name = "neulander-core"
description = "A core library for the NeuLANDER project"
readme = "README.md"
authors = [{ name = "vvcb", email = "vvcb.n1@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.24.1",
    "faststream[rabbit]>=0.5.34",
    "pydantic>=2.10",
    "python-dotenv>=1.0.1",
]
dynamic = ["version"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = 'src/neulander_core/__init__.py'

[dependency-groups]
dev = ["ruff>=0.9.4"]
